refactored sync package into separate files
