--- ibft/sync/history.go
@@ -1,276 +1,0 @@
-package sync
-
-import (
-	"encoding/hex"
-	"fmt"
-	"github.com/bloxapp/ssv/ibft/proto"
-	"github.com/bloxapp/ssv/network"
-	"github.com/bloxapp/ssv/storage/collections"
-	"github.com/bloxapp/ssv/storage/kv"
-	"github.com/pkg/errors"
-	"go.uber.org/zap"
-	"sync"
-	"time"
-)
-
-// HistorySync is responsible for syncing and iBFT instance when needed by
-// fetching decided messages from the network
-type HistorySync struct {
-	logger              *zap.Logger
-	publicKey           []byte
-	network             network.Network
-	ibftStorage         collections.Iibft
-	validateDecidedMsgF func(msg *proto.SignedMessage) error
-	identifier          []byte
-}
-
-// NewHistorySync returns a new instance of HistorySync
-func NewHistorySync(logger *zap.Logger, publicKey []byte, identifier []byte, network network.Network, ibftStorage collections.Iibft, validateDecidedMsgF func(msg *proto.SignedMessage) error) *HistorySync {
-	return &HistorySync{
-		logger:              logger,
-		publicKey:           publicKey,
-		identifier:          identifier,
-		network:             network,
-		validateDecidedMsgF: validateDecidedMsgF,
-		ibftStorage:         ibftStorage,
-	}
-}
-
-// Start the sync
-func (s *HistorySync) Start() error {
-	start := time.Now()
-	// fetch remote highest
-	remoteHighest, fromPeer, err := s.findHighestInstance()
-	if err != nil {
-		return errors.Wrap(err, "could not fetch highest instance during sync")
-	}
-	if remoteHighest == nil { // could not find highest, there isn't one
-		s.logger.Info("node is synced: could not find any peer with highest decided, assuming sequence number is 0", zap.String("duration", time.Since(start).String()))
-		return nil
-	}
-
-	// fetch local highest
-	localHighest, err := s.ibftStorage.GetHighestDecidedInstance(s.identifier)
-	if err != nil && err.Error() != kv.EntryNotFoundError { // if not found, don't continue with sync
-		return errors.Wrap(err, "could not fetch local highest instance during sync")
-	}
-
-	syncStartSeqNumber := uint64(0)
-	if localHighest != nil {
-		syncStartSeqNumber = localHighest.Message.SeqNumber
-	}
-
-	specialStartupCase := err != nil && err.Error() == kv.EntryNotFoundError && remoteHighest.Message.SeqNumber == 0 // in case when remote return seqNum of 0 and local is empty (notFound) we need to save and not assumed to be synced
-	if !specialStartupCase {
-		// check we are behind and need to sync
-		if syncStartSeqNumber >= remoteHighest.Message.SeqNumber {
-			s.logger.Info("node is synced", zap.Uint64("highest seq", syncStartSeqNumber), zap.String("duration", time.Since(start).String()))
-			return nil
-		}
-	}
-
-	// fetch, validate and save missing data
-	highestSaved, err := s.fetchValidateAndSaveInstances(fromPeer, syncStartSeqNumber, remoteHighest.Message.SeqNumber)
-	if err != nil {
-		return errors.Wrap(err, "could not fetch decided by range during sync")
-	}
-
-	// save highest
-	if highestSaved != nil {
-		if err := s.ibftStorage.SaveHighestDecidedInstance(highestSaved); err != nil {
-			return errors.Wrap(err, "could not save highest decided msg during sync")
-		}
-	}
-
-	s.logger.Info("finished syncing", zap.Uint64("highest seq", highestSaved.Message.SeqNumber), zap.String("duration", time.Since(start).String()))
-	return nil
-}
-
-// findHighestInstance returns the highest found decided signed message and the peer it was received from
-func (s *HistorySync) findHighestInstance() (*proto.SignedMessage, string, error) {
-	// pick up to 4 peers
-	// TODO - why 4? should be set as param?
-	// TODO select peers by quality/ score?
-	// TODO - should be changed to support multi duty
-	usedPeers, err := s.network.AllPeers(s.publicKey)
-	if err != nil {
-		return nil, "", err
-	}
-	if len(usedPeers) > 4 {
-		usedPeers = usedPeers[:4]
-	}
-
-	results := s.getHighestDecidedFromPeers(usedPeers)
-
-	// no decided msgs were received from peers, return error
-	if len(results) == 0 {
-		s.logger.Debug("could not fetch highest decided from peers",
-			zap.String("identifier", hex.EncodeToString(s.identifier)))
-		return nil, "", errors.New("could not fetch highest decided from peers")
-	}
-
-	// find the highest decided within the incoming messages
-	var ret *proto.SignedMessage
-	var fromPeer string
-	for _, res := range results {
-		if res.Error == kv.EntryNotFoundError {
-			continue
-		}
-
-		if ret == nil {
-			ret = res.SignedMessages[0]
-			fromPeer = res.FromPeerID
-		}
-		if ret.Message.SeqNumber < res.SignedMessages[0].Message.SeqNumber {
-			ret = res.SignedMessages[0]
-			fromPeer = res.FromPeerID
-		}
-	}
-
-	// highest decided is a nil msg, meaning no decided found from peers. This can happen if no previous decided instance exists.
-	if ret == nil {
-		return nil, "", nil
-	}
-
-	// found a valid highest decided
-	return ret, fromPeer, nil
-}
-
-// getHighestDecidedFromPeers receives highest decided messages from peers
-func (s *HistorySync) getHighestDecidedFromPeers(peers []string) []*network.SyncMessage {
-	var results []*network.SyncMessage
-	var wg sync.WaitGroup
-	var lock sync.Mutex
-
-	// peer's highest decided message will be added to results if:
-	//  1. not-found-error (i.e. no history)
-	//  3. message is valid
-	for i, p := range peers {
-		wg.Add(1)
-		go func(index int, peer string, wg *sync.WaitGroup) {
-			defer wg.Done()
-			res, err := s.network.GetHighestDecidedInstance(peer, &network.SyncMessage{
-				Type:   network.Sync_GetHighestType,
-				Lambda: s.identifier,
-			})
-			if err != nil {
-				s.logger.Error("received error when fetching highest decided", zap.Error(err),
-					zap.String("identifier", hex.EncodeToString(s.identifier)))
-				return
-			}
-			// backwards compatibility for version < v0.0.12, where EntryNotFoundError didn't exist
-			// TODO: can be removed once the release is deprecated
-			if len(res.SignedMessages) == 0 {
-				res.Error = kv.EntryNotFoundError
-			}
-
-			if len(res.Error) > 0 {
-				// assuming not found is a valid scenario (e.g. new validator)
-				// therefore we count the result now, and it will be identified afterwards in findHighestInstance()
-				if res.Error == kv.EntryNotFoundError {
-					lock.Lock()
-					results = append(results, res)
-					lock.Unlock()
-				} else {
-					s.logger.Error("received error when fetching highest decided", zap.Error(err),
-						zap.String("identifier", hex.EncodeToString(s.identifier)))
-				}
-				return
-			}
-
-			if len(res.SignedMessages) != 1 {
-				s.logger.Debug("received multiple signed messages", zap.Error(err),
-					zap.String("identifier", hex.EncodeToString(s.identifier)))
-				return
-			}
-			if err := s.validateDecidedMsgF(res.SignedMessages[0]); err != nil {
-				s.logger.Debug("received invalid highest decided", zap.Error(err),
-					zap.String("identifier", hex.EncodeToString(s.identifier)))
-				return
-			}
-
-			lock.Lock()
-			results = append(results, res)
-			lock.Unlock()
-		}(i, p, &wg)
-	}
-	wg.Wait()
-
-	return results
-}
-
-// FetchValidateAndSaveInstances fetches, validates and saves decided messages from the P2P network.
-// Range is start to end seq including
-func (s *HistorySync) fetchValidateAndSaveInstances(fromPeer string, startSeq uint64, endSeq uint64) (highestSaved *proto.SignedMessage, err error) {
-	failCount := 0
-	start := startSeq
-	done := false
-	var latestError error
-	for {
-		if failCount == 5 {
-			return highestSaved, latestError
-		}
-		if done {
-			return highestSaved, nil
-		}
-
-		res, err := s.network.GetDecidedByRange(fromPeer, &network.SyncMessage{
-			Lambda: s.identifier,
-			Params: []uint64{start, endSeq},
-			Type:   network.Sync_GetInstanceRange,
-		})
-		if err != nil {
-			failCount++
-			latestError = err
-			continue
-		}
-
-		// organize signed msgs into a map where the key is the sequence number
-		// This is for verifying all expected sequence numbers where returned from peer
-		foundSeqs := make(map[uint64]*proto.SignedMessage)
-		for _, msg := range res.SignedMessages {
-			foundSeqs[msg.Message.SeqNumber] = msg
-		}
-
-		msgCount := len(res.SignedMessages)
-		// validate and save
-		for i := start; i <= endSeq; i++ {
-			msg, found := foundSeqs[i]
-			if !found {
-				failCount++
-				latestError = errors.Errorf("returned decided by range messages miss sequence number %d", i)
-				s.logger.Debug("decided by range messages miss sequence number",
-					zap.Uint64("seq", i), zap.Int("msgCount", msgCount))
-				break
-			}
-			// counting all the messages that were visited
-			msgCount--
-			// if msg is invalid, break and try again with an updated start seq
-			if s.validateDecidedMsgF(msg) != nil {
-				start = msg.Message.SeqNumber
-				continue
-			}
-
-			// save
-			if err := s.ibftStorage.SaveDecided(msg); err != nil {
-				return highestSaved, err
-			}
-
-			// set highest
-			if highestSaved == nil || highestSaved.Message.SeqNumber < msg.Message.SeqNumber {
-				highestSaved = msg
-			}
-
-			start = msg.Message.SeqNumber + 1
-
-			if msg.Message.SeqNumber == endSeq {
-				done = true
-			}
-			// if the current messages batch was processed -> break loop and start the next batch
-			if msgCount == 0 {
-				break
-			}
-		}
-		s.logger.Info(fmt.Sprintf("fetched and saved instances up to sequence number %d", start - 1))
-	}
-}
